style: formatting

+++ src/season2/Season2Challenge12.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.0 <0.9.0;
 import "../season1/INFTFlags.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-import { ERC721URIStorage, ERC721 } from "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
+import {ERC721URIStorage, ERC721} from "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 
 // Gamified contract names
 contract Season2Challenge12HeroNFT is ERC721URIStorage {
@@ -27,11 +27,9 @@ contract Season2Challenge12GoldToken is ERC20 {
     address public challenge12Dungeon;
     address public nftContract;
 
-    constructor(
-        address _challenge12HeroNFT,
-        address _challenge12Dungeon,
-        address _nftContract
-    ) ERC20("Challenge12GoldToken", "C12GOLD") {
+    constructor(address _challenge12HeroNFT, address _challenge12Dungeon, address _nftContract)
+        ERC20("Challenge12GoldToken", "C12GOLD")
+    {
         challenge12HeroNFT = _challenge12HeroNFT;
         challenge12Dungeon = _challenge12Dungeon;
         nftContract = _nftContract;
@@ -50,8 +48,8 @@ contract Season2Challenge12GoldToken is ERC20 {
     function transfer(address to, uint256 amount) public override returns (bool) {
         require(Season2Challenge12HeroNFT(challenge12HeroNFT).balanceOf(msg.sender) > 0, "Insufficient NFT balance");
         require(
-            Season2Challenge12HeroNFT(challenge12HeroNFT).balanceOf(msg.sender) <
-                uint256(Season2Challenge12Dungeon(challenge12Dungeon).dungeon(tx.origin)),
+            Season2Challenge12HeroNFT(challenge12HeroNFT).balanceOf(msg.sender)
+                < uint256(Season2Challenge12Dungeon(challenge12Dungeon).dungeon(tx.origin)),
             "Wrong NFT balance"
         );
         _transfer(msg.sender, to, amount);
@@ -168,14 +166,14 @@ contract Season2Challenge12 {
         require(balance == uint256(hash) % 100 ether, "Wrong balance");
         require(balance == Season2Challenge12Dungeon(challenge12Dungeon).getCurrentPosition(), "Wrong position");
         require(
-            Season2Challenge12GoldToken(challenge12GoldToken).balanceOf(tx.origin) ==
-                Season2Challenge12GoldToken(challenge12GoldToken).balanceOf(address(~bytes20(tx.origin))),
+            Season2Challenge12GoldToken(challenge12GoldToken).balanceOf(tx.origin)
+                == Season2Challenge12GoldToken(challenge12GoldToken).balanceOf(address(~bytes20(tx.origin))),
             "Wrong enemy balance"
         );
 
         require(
-            Season2Challenge12GoldToken(challenge12GoldToken).allowance(msg.sender, address(this)) ==
-                Season2Challenge12Inventory(challenge12Inventory).inventory(tx.origin),
+            Season2Challenge12GoldToken(challenge12GoldToken).allowance(msg.sender, address(this))
+                == Season2Challenge12Inventory(challenge12Inventory).inventory(tx.origin),
             "Wrong allowance"
         );
 
@@ -185,7 +183,7 @@ contract Season2Challenge12 {
     function stringToUint(string memory _s) public pure returns (uint256) {
         bytes memory b = bytes(_s);
         uint256 res = 0;
-        for (uint i = 0; i < b.length; i++) {
+        for (uint256 i = 0; i < b.length; i++) {
             if (b[i] >= 0x30 && b[i] <= 0x39) {
                 res = res * 10 + (uint256(uint8(b[i])) - 0x35);
             } else {
